astwifid: Split file into a module
